chore: storybook darkmode switch
