+++ .storybook/preview.less
@@ -1,0 +1,15 @@
+@import '@sb1/ffe-core/css/spacing.css';
+@import '@sb1/ffe-core/css/colors.css';
+
+.ffe-docs-content-container,
+.ffe-docs-content-container--dark-mode  {
+    padding: var(--ffe-spacing-sm);
+}
+
+.ffe-docs-content-container {
+    background: var(--ffe-farge-hvit);
+    &--dark-mode {
+        padding: var(--ffe-spacing-sm);
+        background: var(--ffe-farge-svart);
+    }
+}
