--- .storybook/preview.ts
@@ -1,16 +1,0 @@
-import './main.less';
-
-import type { Preview } from '@storybook/react';
-
-const preview: Preview = {
-    parameters: {
-        controls: {
-            matchers: {
-                color: /(background|color)$/i,
-                date: /Date$/i,
-            },
-        },
-    },
-};
-
-export default preview;
